backend/requirements.txt
# Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
aiosqlite==0.19.0
psycopg2-binary==2.9.9
alembic==1.12.1

# Validation and settings
pydantic==2.5.2
pydantic-settings==2.1.0
email-validator==2.1.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Document Processing
pdfplumber==0.10.3
pytesseract==0.3.10
Pillow==10.1.0

# Data Processing
pandas==2.1.3
numpy==1.26.2

# PDF Generation
reportlab==4.0.7

# Utilities
python-dateutil==2.8.2
aiofiles==23.2.1
